Change Symbol ownership logic
Symbol was previously owned by Scope objects, which were then deleted during the semantic phase. This would have led to dangling Symbol* pointers stored in the AST, making them unusable in the codegen stage. Currently, unique pointers of Symbols are lazily stored inside the symbol table, as this object will not be deleted till the end of the compilation