Rework function and parameter symbols creation
There was a possible critical bug in the way function symbols and parameter symbols were created.
In fact, parameter symbols were created twice in different scenarios in FuncDef::sem (or makeFunctionSymbol):
1. shared_ptr<ParamSymbol> added to the FuncSymbol (the function type signature)
2. unique_ptr<ParamSymbol>, declared live in the SemContext for usage inside a Scope.

This would cause a pointer mismatch, as the symbol in the FuncSymbol and the symbol used by the semantic Scope objects (later stored in the ast to be used in codegen)
would point to different addresses in memory.

Fix: We ensured that ParamSymbols are created once, owned by the SymbolTable, and that raw pointers to these exact same objects are shared by both the FuncSymbol
and the Scope.

Co-authored-by: geokoko <geokoko3@outlook.com.gr>
Co-authored-by: vasilikitsouk <Vasilikitsouk5@gmail.com>